Remove unneccessary requirejs config, organize modules as Require sees them
Not sure if feature detects are better in thier own top-level directory, or in
src where require assumes they will be. Trivial change either way, trying this
out for now.

Making require output a single file removes the dependency on grunt-contrib-copy
as well.

Not sure if we need findNestedDependencies. Output is same with it set to false,
so leaving out for now.